requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
prometheus-fastapi-instrumentator>=6.1.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
